refactor: gate src complexity before test support files
Replace the logic/TUI two-tier split with src at MI 52 first and test/
support files at MI 40 second so the fast verify path scores all source
before slower test fixtures.
