+++ src/tui/board/store.tsx
@@ -1,14 +1,10 @@
 import type { TuiPluginApi } from "@opencode-ai/plugin/tui"
-import { createMemo, createSignal } from "solid-js"
 import type { Event, SessionStatus } from "@opencode-ai/sdk/v2"
 import { commandPlan, configuredScopes } from "../../domain/task/commands"
 import { inProgressCap, sendBackStopThreshold, squashMerge } from "../../domain/task/policy"
-import { getFilter, setFilter as persistFilter } from "../session/preferences"
+import { setFilter as persistFilter } from "../session/preferences"
 import type { ColumnType } from "../../domain/task/types"
 import { ROUTE, type BoardSession } from "../types"
-import type { UpdateStatus } from "../updates/check"
-import { filterSessions, groupCardsByColumn } from "./store/sessions"
-import { createNotices } from "./store/notices"
 import {
   moveDenyReason,
   reorder,
@@ -17,9 +13,9 @@ import {
   selectEdge,
   selectStep,
   selectedSession,
-  type StoreState,
 } from "./store/selection"
 import { deleteSelected, moveByDirection, moveTo, refresh } from "./store/refresh"
+import { createStoreState } from "./store/state-setup"
 
 export function createBoardStore(api: TuiPluginApi, options?: Record<string, unknown>) {
   const squash = squashMerge(options)
@@ -29,48 +25,14 @@ export function createBoardStore(api: TuiPluginApi, options?: Record<string, unk
   const check = checkCommands.map((command) => command.command).join(" && ") || undefined
   const scopes = configuredScopes(options)
   const sendBackThreshold = sendBackStopThreshold(options)
-  const [sessions, setSessions] = createSignal<BoardSession[]>([])
-  const [updateStatus, setUpdateStatus] = createSignal<UpdateStatus>()
-  const [selectedID, setSelectedID] = createSignal<string | undefined>()
-  const [selectedColumn, setSelectedColumn] = createSignal<ColumnType>("backlog")
-  const [filter, setFilterSignal] = createSignal(getFilter(api))
-  const [orders, setOrders] = createSignal<Record<ColumnType, readonly string[]>>({
-    backlog: [],
-    in_progress: [],
-    review: [],
-    done: [],
-  })
-  const filteredSessions = createMemo(() => filterSessions(sessions(), filter()))
-  const columns = createMemo(() => groupCardsByColumn(filteredSessions(), orders()))
-  const { notices, notify, toastError, runWithToast } = createNotices()
-  const [sessionStatuses, setSessionStatuses] = createSignal<Record<string, SessionStatus["type"]>>({})
-
-  const s: StoreState = {
-    api,
-    options,
-    sessions,
-    setSessions,
-    selectedID,
-    setSelectedID,
-    selectedColumn,
-    setSelectedColumn,
-    filter,
-    setFilterSignal,
-    orders,
-    setOrders,
-    columns,
-    notify,
-    toastError,
-    runWithToast,
-    refreshState: { started: 0, completed: 0, helperFailuresSeen: new Map(), awaitingPermissionsSeen: new Set() },
-  }
+  const state = createStoreState(api, options)
 
   return {
-    sessions,
-    selected: selectedID,
-    selectedSession: () => selectedSession(s),
-    selectedColumn,
-    filter,
+    sessions: state.sessions,
+    selected: state.selectedID,
+    selectedSession: () => selectedSession(state.s),
+    selectedColumn: state.selectedColumn,
+    filter: state.filter,
     squashMerge: squash,
     setupCommand: setup,
     checkCommand: check,
@@ -79,32 +41,32 @@ export function createBoardStore(api: TuiPluginApi, options?: Record<string, unk
     configuredScopes: scopes,
     sendBackStopThreshold: sendBackThreshold,
     inProgressCap: inProgressCap(options),
-    columns,
-    notices,
-    notify,
-    updateStatus,
-    setUpdateStatus,
-    select: (column: ColumnType, id: string | undefined) => select(s, column, id),
-    selectNext: () => selectStep(s, 1),
-    selectPrevious: () => selectStep(s, -1),
-    selectNextColumn: () => selectColumnStep(s, 1),
-    selectPrevColumn: () => selectColumnStep(s, -1),
-    selectFirst: () => selectEdge(s, "first"),
-    selectLast: () => selectEdge(s, "last"),
-    reorder: (direction: 1 | -1) => reorder(s, direction),
-    sessionStatus: (sessionID: string): SessionStatus["type"] | undefined => sessionStatuses()[sessionID],
+    columns: state.columns,
+    notices: state.notices,
+    notify: state.notify,
+    updateStatus: state.updateStatus,
+    setUpdateStatus: state.setUpdateStatus,
+    select: (column: ColumnType, id: string | undefined) => select(state.s, column, id),
+    selectNext: () => selectStep(state.s, 1),
+    selectPrevious: () => selectStep(state.s, -1),
+    selectNextColumn: () => selectColumnStep(state.s, 1),
+    selectPrevColumn: () => selectColumnStep(state.s, -1),
+    selectFirst: () => selectEdge(state.s, "first"),
+    selectLast: () => selectEdge(state.s, "last"),
+    reorder: (direction: 1 | -1) => reorder(state.s, direction),
+    sessionStatus: (sessionID: string): SessionStatus["type"] | undefined => state.sessionStatuses()[sessionID],
     setSessionStatus: (sessionID: string, status: SessionStatus["type"]) =>
-      setSessionStatuses((current) => ({ ...current, [sessionID]: status })),
-    moveNext: () => moveByDirection(s, 1),
-    movePrevious: () => moveByDirection(s, -1),
-    moveTo: (status: ColumnType) => moveTo(s, status),
-    moveDenyReason: (status: ColumnType, session: BoardSession) => moveDenyReason(s, status, session),
-    deleteSelected: () => deleteSelected(s),
-    refresh: () => refresh(s),
+      state.setSessionStatuses((current) => ({ ...current, [sessionID]: status })),
+    moveNext: () => moveByDirection(state.s, 1),
+    movePrevious: () => moveByDirection(state.s, -1),
+    moveTo: (status: ColumnType) => moveTo(state.s, status),
+    moveDenyReason: (status: ColumnType, session: BoardSession) => moveDenyReason(state.s, status, session),
+    deleteSelected: () => deleteSelected(state.s),
+    refresh: () => refresh(state.s),
     setFilter(value: string) {
-      setFilterSignal(value)
+      state.setFilterSignal(value)
       persistFilter(api, value)
-      select(s, selectedColumn(), selectedID())
+      select(state.s, state.selectedColumn(), state.selectedID())
     },
   }
 }
